grus: overlay: Use stock-like thinner notch
Took ginkgo's wider notch cutout to start with. [1]

After then iterate with Yann's SVG editor [2], Inkscape
and Mistkeithy's SetWallpaper app [3] on 12.5.1.0 RFBMIXM (Android 11)
to make the notch thinner enough.

In the stock firmware, there is a toggle with title "Adjust node shape"
and with detail as "Turn on to smoothen notch curves".

Current cutout config goes for the thinner one (smoothen is disabled),
but I kept it a few pixels wider than the stock.

For real world photos take a look at Rocky7842's tree! [4]

[1]: crdroidandroid/android_device_xiaomi_ginkgo@c534f7685b7b
[2]: https://yqnn.github.io/svg-path-editor/
[3]: https://github.com/Mistkeithy/SetWallpaper
[4]: Rocky7842/android_device_xiaomi_grus#1

Signed-off-by: Szabolcs Hubai <szab.hu@gmail.com>